feat: How_to_Guides: Add section on Y2K38
Y2K38 issue has been fixed in Scarthgap, but is still present in
Kirkstone, and hence also in Kirkstone images of AM335x and AM437x.
Since 9.3 release is going to be a Kirkstone one, add a How_to_Guide
section of the presence of this issue, and some solutions on how to
deal with it.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>